+++ crates/server/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod api;
-pub mod auth;
 pub mod components;
 pub mod configuration;
 pub mod domain;
