+++ crates/server/src/rpc/auth.rs
